[1/2] change hardware keyboard layout on the fly
This change will allow reconfiguration of the layout of the hardware keyboard.
The UI change is in packages/apps/Settings.
One important drawback is that the change will only be in effect for
activities started _after_ the change. This could partially be circumvented
by forcing all input devices to be reopened, but this does not work for
the built-in keyboards, which is one of the most important use cases for
this. A workaround is to kill any previously started app in the Settings->Apps
dialog or via a task manager.

The early beginnings of this code were inspired by the work of
Christopher-Eyk Hrabia (https://github.com/cehberlin).